ReaderFragment has optional concreteType
Summary:
== Motivation ==

The goal of this stack is to more accurately model GraphQL semantics so that client-side execution matches server-side execution. Specifically, the server has full knowledge of the schema (which types implement which interfaces) and uses this to evaluate or skip type-conditional selections (inline fragments and fragment spreads). Relay doesn't currently have full knowledge of the schema and uses simple rules to decide which type-conditional selections to evaluate, which can result in incorrectly evaluating selections that (per schema/spec) should not be evaluated. There are two cases:

1. Reading fragment spreads where a concrete fragment is spread within an abstract type. Currently Relay cannot distinguish whether the fragment is on a concrete type or not, and always tries to read the fragment. Technically, the fragment shouldn't be evaluated (read) at all.
2. Reading inline fragments.

== This Diff ==

This diff is //part// of addressing case 1. RelayReader should only read a concrete fragment (fragment on a concrete type) if the type of the root object matches the fragment's type. However, we can't implement this today because fragments have a `type: string` which doesn't indicate if the type is concrete or now. **This diff** changes replaces that field with `concreteType: ?string`, where the value is only present if the type is concrete. This diff only changes the compiler, not the runtime. Future diffs will use this knowledge to more precisely evaluate the fragment.

allow-large-files

Reviewed By: jstejada, kassens

Differential Revision: D21281486

fbshipit-source-id: b7109f7ab2309f0aa148a133ac91bb5f6f3755a5