ensure stable compiler output
Summary:
I noticed running the compiler multiple times produces a different diff each time. This makes debugging differences harder.

This diff is fixing that by using insertion order based `IndexMap` in a few places we previously used a randomized `HashMap`.

Eventually, we shouldn't have the need for this anymore as the output shouldn't depend on the input. I imagine this'll be fixed when we achieve parity with JS.

Also switched over to `FnvHashMap` in some places for generally faster perf.

Reviewed By: alunyov

Differential Revision: D21663376

fbshipit-source-id: c73f81e0e2e6ccf6028eab4029f585a6f8dadc37