Fix bug with treatMissingFieldsAsNull flag and client extensions (#3078)
Summary:
I ran into a bug when using `treatMissingFieldsAsNull`, which has to do with the exceptions where a missing field is actually allowed: client extensions and fields on abstract types.

The issue I ran into was with the former, but I also fixed the problem for the latter.

The problem was that when selecting client fields in a fragment that ends up in a query, the normalization process will attempt to write to it with a `null`. This is not right, as we would be resetting what's already in the store for client fields. Moreover, if selecting the `__id` client field, this would even throw an error, as it's not allowed to write null to that field.
Pull Request resolved: https://github.com/facebook/relay/pull/3078

Reviewed By: kassens

Differential Revision: D21622018

Pulled By: josephsavona

fbshipit-source-id: 471a45a55405e327cd8832a6821978a9e88b90a2