Parallelize building projects
Summary:
Parallelizes the transform portion of the projects, leaving the persist and write_artifacts stages sequential.

Before:
* for each project (*sequential*):
  * apply transforms
  * generate artifacts
  * persist
  * update artifacts from persist step
  * write artifacts

After:
  * for all projects (**in parallel**):
    * apply transforms
    * generate artifacts
  * for each project (*sequential*):
    * persist
    * update artifacts from persist step
    * write artifacts

Note that as part of doing this I needed to remove the lifetime from the `Artifact` type, since there was effectively a self-referential pointer from the Artifact result value to the Programs result value. Instead, Artifacts now hold an Arc of their fragments/operations instead of a regular reference.

Reviewed By: alunyov

Differential Revision: D22096415

fbshipit-source-id: e9521905e399056a2ae648832ba48839e2ca225f