move root setting into config
Summary:
The config path and root of the config are tightly coupled. This change makes it
so that the config has an optional `root` field to set the root of the compiler.

By default, the directory of the config is the root, otherwise the relative path
to the root is used.

Reviewed By: josephsavona

Differential Revision: D22498223

fbshipit-source-id: b857dfa77c9f815802a4b96dc07a4111af717aff