rename Definition to TypeSystemDefinition
Summary:
This stack starts to move the schema parsing into the `graphql-syntax` crate. Eventually, they should share one lexer and parser, maybe with variants to parse "schema only documents". One example that we should get rid of the schema parser is that it's error handling is pretty bad and we should have better errors.

This diff renames `schema::Definition` to `schema::TypeSystemDefinition` in preparation to merging into `graphql-syntax`.

Reviewed By: josephsavona

Differential Revision: D22922965

fbshipit-source-id: ef57e409c280431f49e0ec805bdb9297a439853a