rename Document to ExecutableDocument and associated functions
Summary:
This stack starts to move the schema parsing into the `graphql-syntax` crate. Eventually, they should share one lexer and parser, maybe with variants to parse "schema only documents". One example that we should get rid of the schema parser is that it's error handling is pretty bad and we should have better errors.

This diff frees up a bunch of names in `graphql-syntax` in order to make it explicitly about the "executable" definitions (fragments and operations):
- `Document` to `ExecutableDocument` in order to free up the `Document` name for a mixed type document similar to `graphql-js`.
- `parse()` to `parse_executable`
- `Parser::parse_document` to `Parser::parse_executable_document`

Reviewed By: josephsavona

Differential Revision: D22922964

fbshipit-source-id: 2c97e39e8fcf085240399aa531f062ebc6563c09