split up undefined operation variables error
Summary:
Splits up the error into individual errors and uses the updated to create a
Diagnostic.

The reason is that I think each undefined variable is a separate error and
should be reported in an IDE as separate issues. This also allows us to make the
error message more descriptive by including the variable name.

Reviewed By: josephsavona

Differential Revision: D22978840

fbshipit-source-id: de869d13297b649259f697c48eb655615e42336d