fix query_builder with disabled base project
Summary:
If a base project is disabled, we still need to parse the sources in order to compile dependent projects. This is probably something we could optimize by only including enabled projects + base projects, but this is a quick and easy fix.

There'd be more to cleanup if we take the enabled flag into account for the query builder + categorizer.

Reviewed By: tyao1

Differential Revision: D23830191

fbshipit-source-id: 44effc973f3c078043d02a17827cf26af9ebd8c4