delete unnecessary printing
Summary:
I was looking at some perf trace seeing a bunch of printing. Upon taking a quick look if it could be optimized, I realized we can just delete the whole thing.

This was used earlier to replicate the source hash that JS generated, but after some hard to fix inconsistencies we just took the original text and forgot to delete this…

Reviewed By: tyao1

Differential Revision: D24040570

fbshipit-source-id: c8ddb05d010c87c512a4e597b3a8be6b1c733f9e